backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
qdrant-client==1.7.3
pydantic==2.6.0
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
httpx==0.26.0
sqlalchemy==2.0.25
asyncpg==0.29.0
